% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_lambdas.R
\name{opt_lambdas}
\alias{opt_lambdas}
\title{Finding the neighborhood of best lambda_H and lambda_K}
\usage{
opt_lambdas(AFSSEN_CV, diff_between_K, diff_within_H)
}
\arguments{
\item{AFFSEN_CV.}{output of the AFSSEN_CV function}

\item{diff_between_K.}{the threshold for finding the optimum K}

\item{diff_within_H.}{the threshold for finding the optimum H}
}
\description{
This function computes the neighborhood of the best lambda_K and lambda_H.
}
